Move `WindowsServiceLogger` to `logs/windows.go`
