+++ cmd/ops_agent_windows/run_windows.go
@@ -25,6 +25,7 @@ import (
 	"github.com/GoogleCloudPlatform/ops-agent/apps"
 	"github.com/GoogleCloudPlatform/ops-agent/confgenerator"
 	"github.com/GoogleCloudPlatform/ops-agent/internal/healthchecks"
+	"github.com/GoogleCloudPlatform/ops-agent/internal/logs"
 	"golang.org/x/sys/windows/svc"
 	"golang.org/x/sys/windows/svc/debug"
 	"golang.org/x/sys/windows/svc/eventlog"
@@ -145,39 +146,9 @@ func getHealthCheckResults() []healthchecks.HealthCheckResult {
 	return gceHealthChecks.RunAllHealthChecks(logger)
 }
 
-type WindowsServiceLogger struct {
-	srv *service
-}
-
-func (wsl WindowsServiceLogger) Infof(format string, v ...any) {
-	wsl.srv.log.Info(EngineEventID, fmt.Sprintf(format, v...))
-}
-
-func (wsl WindowsServiceLogger) Warnf(format string, v ...any) {
-	wsl.srv.log.Warning(EngineEventID, fmt.Sprintf(format, v...))
-}
-
-func (wsl WindowsServiceLogger) Errorf(format string, v ...any) {
-	wsl.srv.log.Error(EngineEventID, fmt.Sprintf(format, v...))
-}
-
-func (wsl WindowsServiceLogger) Infow(msg string, keysAndValues ...any) {
-	wsl.srv.log.Info(EngineEventID, msg)
-}
-
-func (wsl WindowsServiceLogger) Warnw(msg string, keysAndValues ...any) {
-	wsl.srv.log.Warning(EngineEventID, msg)
-}
-
-func (wsl WindowsServiceLogger) Errorw(msg string, keysAndValues ...any) {
-	wsl.srv.log.Warning(EngineEventID, msg)
-}
-
-func (wsl WindowsServiceLogger) Println(v ...any) {}
-
 func (srv *service) runHealthChecks() {
 	healthCheckResults := getHealthCheckResults()
-	logger := WindowsServiceLogger{srv}
+	logger := logs.WindowsServiceLogger{EventID: EngineEventID, Logger: srv.log}
 	healthchecks.LogHealthCheckResults(healthCheckResults, logger)
 	srv.log.Info(EngineEventID, "Startup checks finished")
 }
